# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MediOrchestra Hospital Management System API
  description: Comprehensive production-ready backend API for MediOrchestra supporting Hospital Admin, Doctor, and OT Manager roles.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api/v1
    description: Local Development Server
paths:
...
  /admin/doctors:
    get:
      security:
        - BearerAuth: []
      summary: List all doctors with filters
      responses:
        '200':
          description: List of doctors
    post:
      security:
        - BearerAuth: []
      summary: Register new doctor and user credentials
      responses:
        '201':
          description: Doctor created
...
  /alerts:
    get:
      security:
        - BearerAuth: []
      summary: List hospital alerts
      responses:
        '200':
          description: Alerts list
    post:
      security:
        - BearerAuth: []
      summary: Create alert
      responses:
        '201':
          description: Alert created
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT